Improve config flow descriptions
This commit updates `strings.json` to provide better descriptions for
data fields requested by `config_flow.py`.

Apparently, this information also has to be duplicated to the
`translations/` directory, even if there are no plans to localize
the component.

A hard refresh of the browser is also often required to make
changes here visible in the UI.