Abstract Carroll Electric API interactions
This commit adds a `CarrollEccHass` class that handles the interface
between Home Assistant and the Carroll Electric website. This class will
handle details such as transforming API output for HomeAssistant and
integrating blocking calls with the HomeAssistant `asyncio` event loops.

This abstraction removes implementation details from the regular
HomeAssistant plugin code and also insulates the browser and
HTTP clients that interact with the Carroll Electric website
from HomeAssistant dependencies and state such that the clients
can be used and tested outside of HomeAssistant.