Add cecc.exceptions
This commit adds custom exceptions to the `cecc` module along with error
handling in the `CarrollEccBrowser` class that makes use of the new
error classes. Pushing the error handling down allows instances of the
Selenium `WebDriverException` to be disambiguated, which eliminates the
need for the separate `test_login()` method.  Eliminating this method
speeds validation up significantly by removing an extra cycle of browser
allocation.