Add client for fetching Carroll Electric data
This commit adds a `CarrollEccAPI` class that uses the `requests`
library to fetch daily usage data from Carroll Electric. Cookies
for a session established through browser login are required to
make API requests, so this commit also updates `CarrollEccBrowser`
with logic for exporting browser cookies to a Python `CookieJar`
instance.

Finally, the `CarrollEccHass` class is updated with logic for
fetching the most recent meter usage available.